refactor(test): extract adapter layer tests to tests/unit/
Move public API tests from 19 adapter source files to client/tests/unit/adapter/.
Delete tests accessing private functions (validate_kfrag_id, validate_binary_data).
Delete repository_impl tests that depended on #[cfg(test)] MockArweaveClient.
Remove all #[cfg(test)] modules from adapter source files.

Deleted tests (private API access):
- ao/message: test_validate_kfrag_id, test_validate_binary_data
- repository_impl/*: all tests (MockArweaveClient dependency)
- arweave/deep_hash: all tests (private functions)
- arweave/merkle: all tests (private functions)
- arweave/transaction: all tests (private functions)
