+++ client/src/adapter/external/arweave/config.rs
@@ -140,40 +140,3 @@ impl Default for ArweaveClientConfig {
         }
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_default_config() {
-        let config = ArweaveClientConfig::default();
-        assert_eq!(config.gateway_url(), DEFAULT_GATEWAY_URL);
-        assert_eq!(config.graphql_url(), DEFAULT_GRAPHQL_URL);
-        assert_eq!(config.timeout_secs(), DEFAULT_TIMEOUT_SECS);
-        assert_eq!(config.max_retries(), DEFAULT_MAX_RETRIES);
-        assert_eq!(config.retry_backoff_ms(), DEFAULT_RETRY_BACKOFF_MS);
-    }
-
-    #[test]
-    fn test_builder_pattern() {
-        let config = ArweaveClientConfig::new()
-            .with_gateway_url("https://custom.gateway.io")
-            .with_graphql_url("https://custom.graphql.io")
-            .with_timeout(60)
-            .with_retries(5, 2000);
-
-        assert_eq!(config.gateway_url(), "https://custom.gateway.io");
-        assert_eq!(config.graphql_url(), "https://custom.graphql.io");
-        assert_eq!(config.timeout_secs(), 60);
-        assert_eq!(config.max_retries(), 5);
-        assert_eq!(config.retry_backoff_ms(), 2000);
-    }
-
-    #[test]
-    fn test_from_env_defaults() {
-        let config = ArweaveClientConfig::from_env().unwrap();
-        assert!(!config.gateway_url().is_empty());
-        assert!(!config.graphql_url().is_empty());
-    }
-}
