+++ client/src/adapter/external/arweave/deep_hash.rs
@@ -159,161 +159,3 @@ pub fn build_signature_message(
 
     Ok(deep_hash(&structure))
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_deep_hash_blob_empty() {
-        let result = deep_hash(&DeepHashItem::Blob(vec![]));
-        assert_eq!(result.len(), 48); // SHA-384 produces 48 bytes
-    }
-
-    #[test]
-    fn test_deep_hash_blob_data() {
-        let result = deep_hash(&DeepHashItem::Blob(b"hello".to_vec()));
-        assert_eq!(result.len(), 48);
-    }
-
-    #[test]
-    fn test_deep_hash_empty_list() {
-        let result = deep_hash(&DeepHashItem::List(vec![]));
-        // Per Arweave spec: empty list produces SHA-384("list0"), NOT treated as empty blob
-        let blob_result = deep_hash(&DeepHashItem::Blob(vec![]));
-        assert_ne!(result, blob_result); // Empty list != empty blob
-        assert_eq!(result.len(), 48); // SHA-384 produces 48 bytes
-    }
-
-    #[test]
-    fn test_deep_hash_single_item_list() {
-        let list_result = deep_hash(&DeepHashItem::List(vec![DeepHashItem::Blob(
-            b"hello".to_vec(),
-        )]));
-        assert_eq!(list_result.len(), 48);
-        // Single item list should NOT equal the blob itself (it's still a list)
-        let blob_result = deep_hash(&DeepHashItem::Blob(b"hello".to_vec()));
-        assert_ne!(list_result, blob_result);
-    }
-
-    #[test]
-    fn test_deep_hash_multiple_items() {
-        let result = deep_hash(&DeepHashItem::List(vec![
-            DeepHashItem::Blob(b"hello".to_vec()),
-            DeepHashItem::Blob(b"world".to_vec()),
-        ]));
-        assert_eq!(result.len(), 48);
-    }
-
-    #[test]
-    fn test_deep_hash_nested_list() {
-        // Test nested structure like tags: [[name, value], [name, value]]
-        let result = deep_hash(&DeepHashItem::List(vec![
-            DeepHashItem::List(vec![
-                DeepHashItem::Blob(b"name1".to_vec()),
-                DeepHashItem::Blob(b"value1".to_vec()),
-            ]),
-            DeepHashItem::List(vec![
-                DeepHashItem::Blob(b"name2".to_vec()),
-                DeepHashItem::Blob(b"value2".to_vec()),
-            ]),
-        ]));
-        assert_eq!(result.len(), 48);
-    }
-
-    #[test]
-    fn test_deep_hash_deterministic() {
-        let input = DeepHashItem::List(vec![
-            DeepHashItem::Blob(b"test".to_vec()),
-            DeepHashItem::Blob(b"content".to_vec()),
-        ]);
-        let result1 = deep_hash(&input);
-        let result2 = deep_hash(&input);
-        assert_eq!(result1, result2);
-    }
-
-    #[test]
-    fn test_build_signature_message_empty_tags() {
-        let result = build_signature_message(
-            b"owner",
-            b"target",
-            "0",
-            "1000",
-            b"last_tx",
-            &[],
-            "100",
-            b"data_root",
-        )
-        .expect("Should succeed with empty tags");
-        assert_eq!(result.len(), 48);
-    }
-
-    #[test]
-    fn test_build_signature_message_with_tags() {
-        use crate::adapter::external::arweave::client::base64url_encode;
-
-        let tags = vec![EncodedTag {
-            name: base64url_encode(b"Content-Type"),
-            value: base64url_encode(b"text/plain"),
-        }];
-
-        let result = build_signature_message(
-            b"owner",
-            b"target",
-            "0",
-            "1000",
-            b"last_tx",
-            &tags,
-            "100",
-            b"data_root",
-        )
-        .expect("Should succeed with valid tags");
-        assert_eq!(result.len(), 48);
-    }
-
-    #[test]
-    fn test_build_signature_message_invalid_tag_name() {
-        let tags = vec![EncodedTag {
-            name: "!!!invalid-base64!!!".to_string(),
-            value: "dGVzdA".to_string(),
-        }];
-
-        let result = build_signature_message(
-            b"owner",
-            b"target",
-            "0",
-            "1000",
-            b"last_tx",
-            &tags,
-            "100",
-            b"data_root",
-        );
-        assert!(result.is_err());
-        let err = result.unwrap_err();
-        assert!(matches!(err, AdapterError::ValidationError { .. }));
-    }
-
-    #[test]
-    fn test_build_signature_message_invalid_tag_value() {
-        use crate::adapter::external::arweave::client::base64url_encode;
-
-        let tags = vec![EncodedTag {
-            name: base64url_encode(b"Content-Type"),
-            value: "!!!invalid-base64!!!".to_string(),
-        }];
-
-        let result = build_signature_message(
-            b"owner",
-            b"target",
-            "0",
-            "1000",
-            b"last_tx",
-            &tags,
-            "100",
-            b"data_root",
-        );
-        assert!(result.is_err());
-        let err = result.unwrap_err();
-        assert!(matches!(err, AdapterError::ValidationError { .. }));
-    }
-}
