+++ client/src/adapter/external/arweave/merkle.rs
@@ -184,70 +184,3 @@ pub fn compute_data_root(payload: &[u8]) -> String {
     let root = build_layers(leaves);
     base64url_encode(&root.id)
 }
-
-#[cfg(test)]
-#[allow(clippy::indexing_slicing)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_chunk_data_empty() {
-        let chunks = chunk_data(&[]);
-        assert!(chunks.is_empty());
-    }
-
-    #[test]
-    fn test_chunk_data_small() {
-        let payload = vec![0u8; 1000];
-        let chunks = chunk_data(&payload);
-        assert_eq!(chunks.len(), 1);
-        assert_eq!(chunks[0].len(), 1000);
-    }
-
-    #[test]
-    fn test_chunk_data_exactly_max() {
-        let payload = vec![0u8; MAX_CHUNK_SIZE];
-        let chunks = chunk_data(&payload);
-        assert_eq!(chunks.len(), 1);
-    }
-
-    #[test]
-    fn test_chunk_data_multiple() {
-        let payload = vec![0u8; MAX_CHUNK_SIZE * 2 + 1000];
-        let chunks = chunk_data(&payload);
-        assert!(chunks.len() >= 2);
-        let total: usize = chunks.iter().map(|c| c.len()).sum();
-        assert_eq!(total, payload.len());
-    }
-
-    #[test]
-    fn test_compute_data_root_empty() {
-        let root = compute_data_root(&[]);
-        assert!(root.is_empty());
-    }
-
-    #[test]
-    fn test_compute_data_root_small() {
-        let payload = b"hello world";
-        let root = compute_data_root(payload);
-        assert!(!root.is_empty());
-        // Base64URL encoded SHA-256 is 43 characters
-        assert!(root.len() >= 40);
-    }
-
-    #[test]
-    fn test_int_to_buffer() {
-        let buf = int_to_buffer(256);
-        assert_eq!(buf.len(), 32);
-        // Last 8 bytes should be big-endian 256
-        assert_eq!(&buf[24..], &[0, 0, 0, 0, 0, 0, 1, 0]);
-    }
-
-    #[test]
-    fn test_deterministic() {
-        let payload = b"test payload for merkle tree";
-        let root1 = compute_data_root(payload);
-        let root2 = compute_data_root(payload);
-        assert_eq!(root1, root2);
-    }
-}
