+++ client/src/adapter/external/arweave/mod.rs
@@ -19,9 +19,6 @@ mod merkle;
 mod transaction;
 mod wallet;
 
-#[cfg(test)]
-mod tests;
-
 pub use client::ArweaveClientImpl;
 pub use config::ArweaveClientConfig;
 pub use wallet::ArweaveWallet;
