+++ client/src/adapter/repository_impl/capsule_impl.rs
@@ -220,110 +220,3 @@ impl<C: ArweaveClient> CapsuleRepository for ArweaveCapsuleRepository<C> {
         Ok(Some(stored.to_entity()))
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::adapter::repository_impl::mock::MockArweaveClient;
-    use crate::repositories::Repository;
-
-    fn create_test_capsule(secret_id: SecretId) -> Capsule {
-        Capsule::new(secret_id, vec![1u8; 100], vec![2u8; 32]).unwrap()
-    }
-
-    #[tokio::test]
-    async fn test_save_and_find_by_id() {
-        let client = MockArweaveClient::new();
-        let repo = ArweaveCapsuleRepository::new(client);
-        let secret_id = SecretId::generate();
-        let capsule = create_test_capsule(secret_id);
-        let id = capsule.id().clone();
-
-        repo.save(&capsule).await.unwrap();
-        let found = repo.find_by_id(&id).await.unwrap();
-
-        assert!(found.is_some());
-        let found = found.unwrap();
-        assert_eq!(found.id(), &id);
-        assert_eq!(found.capsule_data().len(), 100);
-    }
-
-    #[tokio::test]
-    async fn test_find_by_id_not_found() {
-        let client = MockArweaveClient::new();
-        let repo = ArweaveCapsuleRepository::new(client);
-        let id = CapsuleId::generate();
-
-        let found = repo.find_by_id(&id).await.unwrap();
-        assert!(found.is_none());
-    }
-
-    #[tokio::test]
-    async fn test_find_by_secret_id() {
-        let client = MockArweaveClient::new();
-        let repo = ArweaveCapsuleRepository::new(client);
-        let secret_id = SecretId::generate();
-        let capsule = create_test_capsule(secret_id.clone());
-
-        repo.save(&capsule).await.unwrap();
-        let found = repo.find_by_secret_id(&secret_id).await.unwrap();
-
-        assert!(found.is_some());
-        assert_eq!(found.unwrap().secret_id(), &secret_id);
-    }
-
-    #[tokio::test]
-    async fn test_find_by_secret_id_not_found() {
-        let client = MockArweaveClient::new();
-        let repo = ArweaveCapsuleRepository::new(client);
-        let secret_id = SecretId::generate();
-
-        let found = repo.find_by_secret_id(&secret_id).await.unwrap();
-        assert!(found.is_none());
-    }
-
-    #[tokio::test]
-    async fn test_delete() {
-        let client = MockArweaveClient::new();
-        let repo = ArweaveCapsuleRepository::new(client);
-        let secret_id = SecretId::generate();
-        let capsule = create_test_capsule(secret_id);
-        let id = capsule.id().clone();
-
-        repo.save(&capsule).await.unwrap();
-        assert!(repo.exists(&id).await.unwrap());
-
-        repo.delete(&id).await.unwrap();
-        assert!(!repo.exists(&id).await.unwrap());
-    }
-
-    #[tokio::test]
-    async fn test_exists() {
-        let client = MockArweaveClient::new();
-        let repo = ArweaveCapsuleRepository::new(client);
-        let secret_id = SecretId::generate();
-        let capsule = create_test_capsule(secret_id);
-        let id = capsule.id().clone();
-
-        assert!(!repo.exists(&id).await.unwrap());
-        repo.save(&capsule).await.unwrap();
-        assert!(repo.exists(&id).await.unwrap());
-    }
-
-    #[tokio::test]
-    async fn test_binary_data_preserved() {
-        let client = MockArweaveClient::new();
-        let repo = ArweaveCapsuleRepository::new(client);
-        let secret_id = SecretId::generate();
-        let capsule_data = vec![0u8, 1, 2, 255, 128, 64];
-        let owner_pk = vec![10u8; 32];
-        let capsule = Capsule::new(secret_id, capsule_data.clone(), owner_pk.clone()).unwrap();
-        let id = capsule.id().clone();
-
-        repo.save(&capsule).await.unwrap();
-        let found = repo.find_by_id(&id).await.unwrap().unwrap();
-
-        assert_eq!(found.capsule_data(), capsule_data.as_slice());
-        assert_eq!(found.owner_public_key(), owner_pk.as_slice());
-    }
-}
