+++ client/src/adapter/repository_impl/cfrag_impl.rs
@@ -291,153 +291,3 @@ impl<C: ArweaveClient> CFragRepository for ArweaveCFragRepository<C> {
         Ok(cfrags.len())
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::adapter::repository_impl::mock::MockArweaveClient;
-    use crate::repositories::Repository;
-
-    fn create_test_cfrag(secret_id: SecretId, kfrag_id: KFragId, holder_index: u8) -> CFrag {
-        CFrag::new(secret_id, kfrag_id, holder_index, vec![1u8; 100]).unwrap()
-    }
-
-    #[tokio::test]
-    async fn test_save_and_find_by_id() {
-        let client = MockArweaveClient::new();
-        let repo = ArweaveCFragRepository::new(client);
-        let secret_id = SecretId::generate();
-        let kfrag_id = KFragId::generate();
-        let cfrag = create_test_cfrag(secret_id, kfrag_id, 0);
-        let id = cfrag.id().clone();
-
-        repo.save(&cfrag).await.unwrap();
-        let found = repo.find_by_id(&id).await.unwrap();
-
-        assert!(found.is_some());
-        let found = found.unwrap();
-        assert_eq!(found.id(), &id);
-        assert_eq!(found.holder_index(), 0);
-    }
-
-    #[tokio::test]
-    async fn test_find_by_id_not_found() {
-        let client = MockArweaveClient::new();
-        let repo = ArweaveCFragRepository::new(client);
-        let id = CFragId::generate();
-
-        let found = repo.find_by_id(&id).await.unwrap();
-        assert!(found.is_none());
-    }
-
-    #[tokio::test]
-    async fn test_find_by_secret_id() {
-        let client = MockArweaveClient::new();
-        let repo = ArweaveCFragRepository::new(client);
-        let secret_id = SecretId::generate();
-
-        let cfrags: Vec<_> = (0..3)
-            .map(|i| create_test_cfrag(secret_id.clone(), KFragId::generate(), i))
-            .collect();
-
-        for cfrag in &cfrags {
-            repo.save(cfrag).await.unwrap();
-        }
-
-        let found = repo.find_by_secret_id(&secret_id).await.unwrap();
-        assert_eq!(found.len(), 3);
-    }
-
-    #[tokio::test]
-    async fn test_find_by_kfrag_id() {
-        let client = MockArweaveClient::new();
-        let repo = ArweaveCFragRepository::new(client);
-        let secret_id = SecretId::generate();
-        let kfrag_id = KFragId::generate();
-
-        let cfrag = create_test_cfrag(secret_id, kfrag_id.clone(), 0);
-        repo.save(&cfrag).await.unwrap();
-
-        let found = repo.find_by_kfrag_id(&kfrag_id).await.unwrap();
-        assert!(found.is_some());
-        assert_eq!(found.unwrap().kfrag_id(), &kfrag_id);
-
-        let not_found = repo.find_by_kfrag_id(&KFragId::generate()).await.unwrap();
-        assert!(not_found.is_none());
-    }
-
-    #[tokio::test]
-    async fn test_delete() {
-        let client = MockArweaveClient::new();
-        let repo = ArweaveCFragRepository::new(client);
-        let secret_id = SecretId::generate();
-        let kfrag_id = KFragId::generate();
-        let cfrag = create_test_cfrag(secret_id, kfrag_id, 0);
-        let id = cfrag.id().clone();
-
-        repo.save(&cfrag).await.unwrap();
-        assert!(repo.exists(&id).await.unwrap());
-
-        repo.delete(&id).await.unwrap();
-        assert!(!repo.exists(&id).await.unwrap());
-    }
-
-    #[tokio::test]
-    async fn test_delete_by_secret_id() {
-        let client = MockArweaveClient::new();
-        let repo = ArweaveCFragRepository::new(client);
-        let secret_id = SecretId::generate();
-        let other_secret_id = SecretId::generate();
-
-        let cfrags: Vec<_> = (0..3)
-            .map(|i| create_test_cfrag(secret_id.clone(), KFragId::generate(), i))
-            .collect();
-        let other_cfrag = create_test_cfrag(other_secret_id.clone(), KFragId::generate(), 0);
-
-        for cfrag in &cfrags {
-            repo.save(cfrag).await.unwrap();
-        }
-        repo.save(&other_cfrag).await.unwrap();
-
-        repo.delete_by_secret_id(&secret_id).await.unwrap();
-
-        let remaining = repo.find_by_secret_id(&secret_id).await.unwrap();
-        assert!(remaining.is_empty());
-
-        let other_remaining = repo.find_by_secret_id(&other_secret_id).await.unwrap();
-        assert_eq!(other_remaining.len(), 1);
-    }
-
-    #[tokio::test]
-    async fn test_count_by_secret_id() {
-        let client = MockArweaveClient::new();
-        let repo = ArweaveCFragRepository::new(client);
-        let secret_id = SecretId::generate();
-
-        assert_eq!(repo.count_by_secret_id(&secret_id).await.unwrap(), 0);
-
-        let cfrags: Vec<_> = (0..3)
-            .map(|i| create_test_cfrag(secret_id.clone(), KFragId::generate(), i))
-            .collect();
-
-        for cfrag in &cfrags {
-            repo.save(cfrag).await.unwrap();
-        }
-
-        assert_eq!(repo.count_by_secret_id(&secret_id).await.unwrap(), 3);
-    }
-
-    #[tokio::test]
-    async fn test_exists() {
-        let client = MockArweaveClient::new();
-        let repo = ArweaveCFragRepository::new(client);
-        let secret_id = SecretId::generate();
-        let kfrag_id = KFragId::generate();
-        let cfrag = create_test_cfrag(secret_id, kfrag_id, 0);
-        let id = cfrag.id().clone();
-
-        assert!(!repo.exists(&id).await.unwrap());
-        repo.save(&cfrag).await.unwrap();
-        assert!(repo.exists(&id).await.unwrap());
-    }
-}
