+++ client/src/adapter/repository_impl/kfrag_impl.rs
@@ -289,130 +289,3 @@ impl<C: ArweaveClient> KFragRepository for ArweaveKFragRepository<C> {
         Ok(())
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::adapter::repository_impl::mock::MockArweaveClient;
-    use crate::repositories::Repository;
-
-    fn create_test_kfrag(secret_id: SecretId, holder_index: u8) -> KFrag {
-        KFrag::new(secret_id, holder_index, 5, vec![1u8; 100]).unwrap()
-    }
-
-    #[tokio::test]
-    async fn test_save_and_find_by_id() {
-        let client = MockArweaveClient::new();
-        let repo = ArweaveKFragRepository::new(client);
-        let secret_id = SecretId::generate();
-        let kfrag = create_test_kfrag(secret_id, 1);
-        let id = kfrag.id().clone();
-
-        repo.save(&kfrag).await.unwrap();
-        let found = repo.find_by_id(&id).await.unwrap();
-
-        assert!(found.is_some());
-        let found = found.unwrap();
-        assert_eq!(found.id(), &id);
-        assert_eq!(found.holder_index(), 1);
-    }
-
-    #[tokio::test]
-    async fn test_find_by_id_not_found() {
-        let client = MockArweaveClient::new();
-        let repo = ArweaveKFragRepository::new(client);
-        let id = KFragId::generate();
-
-        let found = repo.find_by_id(&id).await.unwrap();
-        assert!(found.is_none());
-    }
-
-    #[tokio::test]
-    async fn test_find_by_secret_id() {
-        let client = MockArweaveClient::new();
-        let repo = ArweaveKFragRepository::new(client);
-        let secret_id = SecretId::generate();
-
-        let kfrags: Vec<_> = (1..=3)
-            .map(|i| create_test_kfrag(secret_id.clone(), i))
-            .collect();
-
-        for kfrag in &kfrags {
-            repo.save(kfrag).await.unwrap();
-        }
-
-        let found = repo.find_by_secret_id(&secret_id).await.unwrap();
-        assert_eq!(found.len(), 3);
-    }
-
-    #[tokio::test]
-    async fn test_find_by_holder_index() {
-        let client = MockArweaveClient::new();
-        let repo = ArweaveKFragRepository::new(client);
-        let secret_id = SecretId::generate();
-
-        let kfrag = create_test_kfrag(secret_id.clone(), 2);
-        repo.save(&kfrag).await.unwrap();
-
-        let found = repo.find_by_holder_index(&secret_id, 2).await.unwrap();
-        assert!(found.is_some());
-        assert_eq!(found.unwrap().holder_index(), 2);
-
-        let not_found = repo.find_by_holder_index(&secret_id, 99).await.unwrap();
-        assert!(not_found.is_none());
-    }
-
-    #[tokio::test]
-    async fn test_delete() {
-        let client = MockArweaveClient::new();
-        let repo = ArweaveKFragRepository::new(client);
-        let secret_id = SecretId::generate();
-        let kfrag = create_test_kfrag(secret_id, 1);
-        let id = kfrag.id().clone();
-
-        repo.save(&kfrag).await.unwrap();
-        assert!(repo.exists(&id).await.unwrap());
-
-        repo.delete(&id).await.unwrap();
-        assert!(!repo.exists(&id).await.unwrap());
-    }
-
-    #[tokio::test]
-    async fn test_delete_by_secret_id() {
-        let client = MockArweaveClient::new();
-        let repo = ArweaveKFragRepository::new(client);
-        let secret_id = SecretId::generate();
-        let other_secret_id = SecretId::generate();
-
-        let kfrags: Vec<_> = (1..=3)
-            .map(|i| create_test_kfrag(secret_id.clone(), i))
-            .collect();
-        let other_kfrag = create_test_kfrag(other_secret_id.clone(), 1);
-
-        for kfrag in &kfrags {
-            repo.save(kfrag).await.unwrap();
-        }
-        repo.save(&other_kfrag).await.unwrap();
-
-        repo.delete_by_secret_id(&secret_id).await.unwrap();
-
-        let remaining = repo.find_by_secret_id(&secret_id).await.unwrap();
-        assert!(remaining.is_empty());
-
-        let other_remaining = repo.find_by_secret_id(&other_secret_id).await.unwrap();
-        assert_eq!(other_remaining.len(), 1);
-    }
-
-    #[tokio::test]
-    async fn test_exists() {
-        let client = MockArweaveClient::new();
-        let repo = ArweaveKFragRepository::new(client);
-        let secret_id = SecretId::generate();
-        let kfrag = create_test_kfrag(secret_id, 1);
-        let id = kfrag.id().clone();
-
-        assert!(!repo.exists(&id).await.unwrap());
-        repo.save(&kfrag).await.unwrap();
-        assert!(repo.exists(&id).await.unwrap());
-    }
-}
