+++ client/src/adapter/repository_impl/mod.rs
@@ -169,188 +169,3 @@ pub mod tag_helpers {
         Tag::new(tag_names::KFRAG_ID, kfrag_id)
     }
 }
-
-/// Mock Arweave client module for testing
-#[cfg(test)]
-pub mod mock {
-    use super::{AdapterError, ArweaveClient, Tag};
-    use async_trait::async_trait;
-    use std::collections::HashMap;
-    use std::sync::RwLock;
-    use std::sync::atomic::{AtomicU64, Ordering};
-
-    /// Entry stored in MockArweaveClient
-    struct StoredEntry {
-        data: Vec<u8>,
-        tags: Vec<Tag>,
-    }
-
-    /// Mock implementation of ArweaveClient for testing
-    ///
-    /// Uses in-memory HashMap storage with RwLock for thread safety.
-    /// Supports tag-based query filtering.
-    pub struct MockArweaveClient {
-        storage: RwLock<HashMap<String, StoredEntry>>,
-        tx_counter: AtomicU64,
-    }
-
-    impl MockArweaveClient {
-        /// Create a new empty MockArweaveClient
-        pub fn new() -> Self {
-            Self {
-                storage: RwLock::new(HashMap::new()),
-                tx_counter: AtomicU64::new(0),
-            }
-        }
-
-        /// Generate a unique transaction ID
-        fn generate_tx_id(&self) -> String {
-            let id = self.tx_counter.fetch_add(1, Ordering::SeqCst);
-            format!("mock_tx_{id}")
-        }
-
-        /// Check if an entry's tags match all query tags
-        fn tags_match(entry_tags: &[Tag], query_tags: &[Tag]) -> bool {
-            query_tags.iter().all(|query_tag| {
-                entry_tags.iter().any(|entry_tag| {
-                    entry_tag.name == query_tag.name && entry_tag.value == query_tag.value
-                })
-            })
-        }
-
-        /// Clear all stored data (useful for test cleanup)
-        pub fn clear(&self) {
-            let mut storage = self.storage.write().unwrap();
-            storage.clear();
-        }
-
-        /// Get the number of stored entries
-        pub fn len(&self) -> usize {
-            let storage = self.storage.read().unwrap();
-            storage.len()
-        }
-
-        /// Check if storage is empty
-        pub fn is_empty(&self) -> bool {
-            self.len() == 0
-        }
-    }
-
-    impl Default for MockArweaveClient {
-        fn default() -> Self {
-            Self::new()
-        }
-    }
-
-    #[async_trait]
-    impl ArweaveClient for MockArweaveClient {
-        async fn get(&self, tx_id: &str) -> Result<Option<Vec<u8>>, AdapterError> {
-            let storage = self.storage.read().unwrap();
-            Ok(storage.get(tx_id).map(|entry| entry.data.clone()))
-        }
-
-        async fn post(&self, data: &[u8], tags: Vec<Tag>) -> Result<String, AdapterError> {
-            let tx_id = self.generate_tx_id();
-            let entry = StoredEntry {
-                data: data.to_vec(),
-                tags,
-            };
-
-            let mut storage = self.storage.write().unwrap();
-            storage.insert(tx_id.clone(), entry);
-
-            Ok(tx_id)
-        }
-
-        async fn query(&self, tags: Vec<Tag>) -> Result<Vec<String>, AdapterError> {
-            let storage = self.storage.read().unwrap();
-            let mut matching_ids: Vec<String> = storage
-                .iter()
-                .filter(|(_, entry)| Self::tags_match(&entry.tags, &tags))
-                .map(|(tx_id, _)| tx_id.clone())
-                .collect();
-
-            // Sort by numeric suffix to ensure proper ordering (mock_tx_0, mock_tx_1, ...)
-            matching_ids.sort_by(|a, b| {
-                let extract_num = |s: &str| {
-                    s.strip_prefix("mock_tx_")
-                        .and_then(|n| n.parse::<u64>().ok())
-                        .unwrap_or(0)
-                };
-                extract_num(a).cmp(&extract_num(b))
-            });
-
-            Ok(matching_ids)
-        }
-    }
-
-    #[cfg(test)]
-    mod tests {
-        use super::*;
-
-        #[tokio::test]
-        async fn test_mock_client_post_and_get() {
-            let client = MockArweaveClient::new();
-            let test_data = vec![1u8, 2, 3, 4, 5];
-            let tags = vec![Tag::new("key", "value")];
-
-            let tx_id = client.post(&test_data, tags).await.unwrap();
-            let retrieved = client.get(&tx_id).await.unwrap();
-
-            assert!(retrieved.is_some());
-            assert_eq!(retrieved.unwrap(), test_data);
-        }
-
-        #[tokio::test]
-        async fn test_mock_client_get_not_found() {
-            let client = MockArweaveClient::new();
-            let result = client.get("nonexistent_tx").await.unwrap();
-            assert!(result.is_none());
-        }
-
-        #[tokio::test]
-        async fn test_mock_client_query_by_tags() {
-            let client = MockArweaveClient::new();
-
-            let tags1 = vec![Tag::new("type", "Secret"), Tag::new("id", "secret1")];
-            let tags2 = vec![Tag::new("type", "Secret"), Tag::new("id", "secret2")];
-            let tags3 = vec![Tag::new("type", "Capsule"), Tag::new("id", "capsule1")];
-
-            client.post(&[1], tags1).await.unwrap();
-            client.post(&[2], tags2).await.unwrap();
-            client.post(&[3], tags3).await.unwrap();
-
-            let query_tags = vec![Tag::new("type", "Secret")];
-            let results = client.query(query_tags).await.unwrap();
-
-            assert_eq!(results.len(), 2);
-        }
-
-        #[tokio::test]
-        async fn test_mock_client_query_no_match() {
-            let client = MockArweaveClient::new();
-
-            let tags = vec![Tag::new("type", "Secret")];
-            client.post(&[1], tags).await.unwrap();
-
-            let query_tags = vec![Tag::new("type", "Capsule")];
-            let results = client.query(query_tags).await.unwrap();
-
-            assert!(results.is_empty());
-        }
-
-        #[tokio::test]
-        async fn test_mock_client_clear() {
-            let client = MockArweaveClient::new();
-
-            client.post(&[1], vec![]).await.unwrap();
-            client.post(&[2], vec![]).await.unwrap();
-
-            assert_eq!(client.len(), 2);
-
-            client.clear();
-
-            assert!(client.is_empty());
-        }
-    }
-}
