+++ client/src/adapter/repository_impl/secret_impl.rs
@@ -229,111 +229,3 @@ impl<C: ArweaveClient> Repository<Secret, SecretId> for ArweaveSecretRepository<
 #[cfg_attr(target_arch = "wasm32", async_trait(?Send))]
 #[cfg_attr(not(target_arch = "wasm32"), async_trait)]
 impl<C: ArweaveClient> SecretRepository for ArweaveSecretRepository<C> {}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::adapter::repository_impl::mock::MockArweaveClient;
-    use crate::repositories::Repository;
-
-    fn create_test_secret() -> Secret {
-        Secret::new(3, 5, vec![1u8; 32]).unwrap()
-    }
-
-    #[tokio::test]
-    async fn test_save_and_find_by_id() {
-        let client = MockArweaveClient::new();
-        let repo = ArweaveSecretRepository::new(client);
-        let secret = create_test_secret();
-        let id = secret.id().clone();
-
-        repo.save(&secret).await.unwrap();
-        let found = repo.find_by_id(&id).await.unwrap();
-
-        assert!(found.is_some());
-        let found = found.unwrap();
-        assert_eq!(found.id(), &id);
-        assert_eq!(found.threshold_k(), 3);
-        assert_eq!(found.threshold_n(), 5);
-    }
-
-    #[tokio::test]
-    async fn test_find_by_id_not_found() {
-        let client = MockArweaveClient::new();
-        let repo = ArweaveSecretRepository::new(client);
-        let id = SecretId::generate();
-
-        let found = repo.find_by_id(&id).await.unwrap();
-        assert!(found.is_none());
-    }
-
-    #[tokio::test]
-    async fn test_exists() {
-        let client = MockArweaveClient::new();
-        let repo = ArweaveSecretRepository::new(client);
-        let secret = create_test_secret();
-        let id = secret.id().clone();
-
-        assert!(!repo.exists(&id).await.unwrap());
-        repo.save(&secret).await.unwrap();
-        assert!(repo.exists(&id).await.unwrap());
-    }
-
-    #[tokio::test]
-    async fn test_delete() {
-        let client = MockArweaveClient::new();
-        let repo = ArweaveSecretRepository::new(client);
-        let secret = create_test_secret();
-        let id = secret.id().clone();
-
-        repo.save(&secret).await.unwrap();
-        assert!(repo.exists(&id).await.unwrap());
-
-        repo.delete(&id).await.unwrap();
-        assert!(!repo.exists(&id).await.unwrap());
-    }
-
-    #[tokio::test]
-    async fn test_delete_nonexistent() {
-        let client = MockArweaveClient::new();
-        let repo = ArweaveSecretRepository::new(client);
-        let id = SecretId::generate();
-
-        let result = repo.delete(&id).await;
-        assert!(result.is_ok());
-    }
-
-    #[tokio::test]
-    async fn test_find_by_ids() {
-        let client = MockArweaveClient::new();
-        let repo = ArweaveSecretRepository::new(client);
-
-        let secrets: Vec<_> = (0..3).map(|_| create_test_secret()).collect();
-        let ids: Vec<_> = secrets.iter().map(|s| s.id().clone()).collect();
-
-        for secret in &secrets {
-            repo.save(secret).await.unwrap();
-        }
-
-        let found = repo.find_by_ids(&ids[0..2]).await.unwrap();
-        assert_eq!(found.len(), 2);
-    }
-
-    #[tokio::test]
-    async fn test_find_by_ids_partial() {
-        let client = MockArweaveClient::new();
-        let repo = ArweaveSecretRepository::new(client);
-
-        let secret = create_test_secret();
-        let existing_id = secret.id().clone();
-        let nonexistent_id = SecretId::generate();
-
-        repo.save(&secret).await.unwrap();
-
-        let found = repo
-            .find_by_ids(&[existing_id, nonexistent_id])
-            .await
-            .unwrap();
-        assert_eq!(found.len(), 1);
-    }
-}
