+++ client/tests/unit/adapter/external/ao/mod.rs
@@ -1,0 +1,6 @@
+mod config;
+#[cfg(feature = "production-ao")]
+mod data_item;
+mod message;
+#[cfg(feature = "production-ao")]
+mod production_client;
