+++ client/tests/unit/adapter/external/arweave/mod.rs
@@ -1,0 +1,3 @@
+mod config;
+mod tests;
+mod wallet;
