+++ client/tests/unit/adapter/external/mock_ao/mod.rs
@@ -1,0 +1,1 @@
+mod client;
