+++ client/tests/unit/adapter/external/mod.rs
@@ -1,0 +1,3 @@
+mod ao;
+mod arweave;
+mod mock_ao;
