+++ client/tests/unit/adapter/mod.rs
@@ -1,0 +1,2 @@
+mod errors;
+mod external;
